+++ test/base/flop/validation_test.exs
@@ -29,10 +29,16 @@ defmodule Flop.ValidationTest do
 
     @derive {Flop.Schema,
              filterable: [],
-             sortable: [:name, :full_name, :thing_count],
+             sortable: [:name, :full_name, :thing_count, :name_length],
              adapter_opts: [
                compound_fields: [full_name: [:family_name, :given_name]],
-               alias_fields: [:thing_count]
+               alias_fields: [:thing_count],
+               custom_fields: [
+                 name_length: [
+                   field_dynamic: {__MODULE__, :name_length, []},
+                   ecto_type: :integer
+                 ]
+               ]
              ]}
 
     schema "things" do
@@ -764,12 +770,19 @@ defmodule Flop.ValidationTest do
       assert {:error, changeset} = validate(params, for: Thing)
 
       assert errors_on(changeset)[:order_by] == [
-               "cursor pagination is not supported for compound and alias fields"
+               "cursor pagination is not supported for compound, alias and" <>
+                 " custom fields"
              ]
     end
 
     test "rejects an alias field as cursor order field" do
       params = %{first: 2, after: @cursor, order_by: [:thing_count]}
+      assert {:error, changeset} = validate(params, for: Thing)
+      assert errors_on(changeset)[:order_by] != nil
+    end
+
+    test "rejects a custom field as cursor order field" do
+      params = %{first: 2, after: @cursor, order_by: [:name_length]}
       assert {:error, changeset} = validate(params, for: Thing)
       assert errors_on(changeset)[:order_by] != nil
     end